backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
sqlalchemy==2.0.41
psycopg[binary]==3.2.13
alembic==1.16.4
pydantic-settings==2.10.1
celery[redis]==5.5.3
redis==5.2.1
httpx==0.28.1
requests==2.32.3
faster-whisper==1.1.1
python-jose[cryptography]==3.5.0
structlog==25.4.0
langdetect==1.0.9
tenacity==9.1.2
openai==1.98.0
